升级为4.0.7,添加交易验签功能

+++ src/main/java/io/hpb/web3/crypto/SignatureDataOperations.java
@@ -12,7 +12,7 @@ public interface SignatureDataOperations {
 
     Sign.SignatureData getSignatureData();
 
-    byte[] getEncodedTransaction(Integer chainId);
+    byte[] getEncodedTransaction(Long chainId);
 
     default String getFrom() throws SignatureException {
         byte[] encodedTransaction = getEncodedTransaction(getChainId());
@@ -44,13 +44,13 @@ public interface SignatureDataOperations {
         return (byte) (realV + inc);
     }
 
-    default Integer getChainId() {
+    default Long getChainId() {
         BigInteger bv = Numeric.toBigInt(getSignatureData().getV());
         long v = bv.longValue();
         if (v == LOWER_REAL_V || v == (LOWER_REAL_V + 1)) {
             return null;
         }
-        Integer chainId = (int)((v - CHAIN_ID_INC) / 2);
+        long chainId = (v - CHAIN_ID_INC) / 2;
         return chainId;
     }
 }
