+++ src/main/java/io/hpb/web3/crypto/SignedRawTransaction.java
@@ -18,7 +18,7 @@ public class SignedRawTransaction extends RawTransaction implements SignatureDat
     }
 
     @Override
-    public byte[] getEncodedTransaction(Integer chainId) {
+    public byte[] getEncodedTransaction(Long chainId) {
         if (null == chainId) {
             return TransactionEncoder.encode(this);
         } else {
